move model to osf model
